+++ src/firecrest/filesystem/transfer/models.py
@@ -14,10 +14,32 @@ class PostFileUploadRequest(FilesystemRequestBase):
     file_name: str
     account: Optional[str] = None
     file_size: int
+    model_config = {
+        "json_schema_extra": {
+            "examples": [
+                {
+                    "path": "/home/user/dir/file",
+                    "file_name": "/path/local/file",
+                    "account": "group",
+                    "file_size": "7340032"
+                }
+            ]
+        }
+    }
 
 
 class PostFileDownloadRequest(FilesystemRequestBase):
     account: Optional[str] = None
+    model_config = {
+        "json_schema_extra": {
+            "examples": [
+                {
+                    "path": "/home/user/dir/file",
+                    "account": "group"
+                }
+            ]
+        }
+    }
 
 
 class PostXferInternalOperationApiResponse(CamelModel):
@@ -51,6 +73,17 @@ class DownloadFileResponse(CamelModel):
 class CopyRequest(FilesystemRequestBase):
     target_path: str
     account: Optional[str] = None
+    model_config = {
+        "json_schema_extra": {
+            "examples": [
+                {
+                    "source_path": "/home/user/dir/file.orig",
+                    "target_path": "/home/user/dir/file.new",
+                    "account": "group"
+                }
+            ]
+        }
+    }
 
 
 class CopyResponse(CamelModel):
@@ -64,6 +97,17 @@ class DeleteResponse(CamelModel):
 class MoveRequest(FilesystemRequestBase):
     target_path: str
     account: Optional[str] = None
+    model_config = {
+        "json_schema_extra": {
+            "examples": [
+                {
+                    "source_path": "/home/user/dir/file.orig",
+                    "target_path": "/home/user/dir/file.new",
+                    "account": "group"
+                }
+            ]
+        }
+    }
 
 
 class MoveResponse(CamelModel):
@@ -75,6 +119,19 @@ class CompressRequest(FilesystemRequestBase):
     account: Optional[str] = None
     match_pattern: Optional[str] = None
     dereference: Optional[bool] = False
+    model_config = {
+        "json_schema_extra": {
+            "examples": [
+                {
+                    "path": "/home/user/dir",
+                    "target_path": "/home/user/file.tar.gz",
+                    "match_pattern": "*./[ab].*\\.txt",
+                    "dereference": "true",
+                    "account": "group"
+                }
+            ]
+        }
+    }
 
 
 class CompressResponse(CamelModel):
@@ -84,6 +141,17 @@ class CompressResponse(CamelModel):
 class ExtractRequest(FilesystemRequestBase):
     target_path: str
     account: Optional[str] = None
+    model_config = {
+        "json_schema_extra": {
+            "examples": [
+                {
+                    "source_path": "/home/user/dir/file.tar.gz",
+                    "target_path": "/home/user/dir",
+                    "account": "group"
+                }
+            ]
+        }
+    }
 
 
 class ExtractResponse(CamelModel):